♻️ :: vote list package refactoring
