+++ app/src/main/java/com/example/dmsport_android/feature/vote/repository/VoteListRepository.kt
@@ -1,6 +1,6 @@
 package com.example.dmsport_android.feature.vote.repository
 
-import com.example.dmsport_android.dto.response.VoteListResponse
+import com.example.dmsport_android.feature.vote.model.VoteListResponse
 import com.example.dmsport_android.network.voteApi
 import com.example.dmsport_android.util.ACCESS_TOKEN
 import retrofit2.Response
